Gen refactor.
